[refactor] rename emitter.js => kuzzleEventEmitter.js
