+++ lib/kuzzle/events/kuzzleEventEmitter.js
