+++ lib/kuzzle/kuzzle.js
@@ -21,7 +21,7 @@
 
 'use strict';
 
-const KuzzleEventEmitter = require('./events/emitter');
+const KuzzleEventEmitter = require('./events/kuzzleEventEmitter');
 const path = require('path');
 const { murmurHash128: murmur } = require('murmurhash-native');
 const stringify = require('json-stable-stringify');
